refactor : Participants 객체를 BlackJackGame으로 변환
